requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-core==1.30.0
azure-ai-formrecognizer==3.2.0
azure-storage-blob==12.24.0

langchain 
langchain-community
unstructured
langchain-openai 